# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API untuk Tubes Pemrograman",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/project": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "List tools and ongoing project",
                "description": "Display a listing of the resource.",
                "operationId": "1a149f32d40ca0b3cb5f2b00b78d901b",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "materials": {
                                            "description": "Tool materials",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ToolMaterial"
                                            }
                                        },
                                        "projects": {
                                            "description": "Tool products",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MachiningProject"
                                            }
                                        },
                                        "machining": {
                                            "description": "Machining parameters",
                                            "properties": {
                                                "workpiece_materials": {
                                                    "description": "Workpiece materials",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "machining_processes": {
                                                    "description": "Machining processes",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "n": {
                                                    "description": "n Constant",
                                                    "type": "number"
                                                },
                                                "c": {
                                                    "description": "c Constant",
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Create tool product",
                "description": "Store a newly created resource in storage.",
                "operationId": "bc8eb8624844672b2cc36d9f12bc547a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMachiningProjectRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMachiningProjectRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMachiningProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MachiningProject"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/project/ongoing": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Create work on an ongoing project",
                "description": "Store a newly created work on an ongoing project in storage.",
                "operationId": "dbc4f1745f6a5eb6cb6bddf442282030",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMachiningProjectWorkRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMachiningProjectWorkRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMachiningProjectWorkRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MachiningProjectWork"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelNotFoundException"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/project/tool/history": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Get history of machining projects",
                "description": "Display the specified history of machining projects.",
                "operationId": "efa6d20a5a5e752bf19114f9b4e95576",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RetreiveToolHistoryRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/RetreiveToolHistoryRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/RetreiveToolHistoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MachiningProject"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelNotFoundException"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ModelNotFoundException": {
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "No query results for model [App\\Models\\ToolMaterial] 1"
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The name field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
            "UnprocessableContentException": {
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Wrong username or password"
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The name field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "RetreiveToolHistoryRequest": {
                "required": [
                    "tool_material_id",
                    "tool_product_id",
                    "tool_item_id"
                ],
                "properties": {
                    "tool_material_id": {
                        "description": "Tool material ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "tool_product_id": {
                        "description": "Tool product ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "tool_item_id": {
                        "description": "Tool item ID",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "StoreMachiningProjectRequest": {
                "required": [
                    "tool_material_id",
                    "tool_product_id",
                    "tool_item_id",
                    "workpiece_material",
                    "machining_process",
                    "cutting_speed",
                    "depth_of_cut",
                    "feeding",
                    "early_tool_life"
                ],
                "properties": {
                    "tool_material_id": {
                        "description": "Tool material ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "tool_product_id": {
                        "description": "Tool product ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "tool_item_id": {
                        "description": "Tool item ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "workpiece_material": {
                        "description": "Workpiece material",
                        "type": "string",
                        "example": "Stainless steel"
                    },
                    "machining_process": {
                        "description": "Machining process",
                        "type": "string",
                        "example": "Milling"
                    },
                    "cutting_speed": {
                        "description": "Cutting speed (m/min)",
                        "type": "integer",
                        "example": "100"
                    },
                    "depth_of_cut": {
                        "description": "Depth of cut (mm)",
                        "type": "integer",
                        "example": "1"
                    },
                    "feeding": {
                        "description": "Feeding (mm/min)",
                        "type": "integer",
                        "example": "100"
                    },
                    "early_tool_life": {
                        "description": "Early tool life (min)",
                        "type": "integer",
                        "example": "10"
                    },
                    "is_active": {
                        "description": "Is active",
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
            "StoreMachiningProjectWorkRequest": {
                "required": [
                    "machining_project_id",
                    "product_id",
                    "initial_diameter",
                    "final_diameter",
                    "workpart_length",
                    "product_quantity",
                    "machining_time"
                ],
                "properties": {
                    "machining_project_id": {
                        "description": "Machining project ID",
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "description": "Product ID",
                        "type": "string",
                        "example": "P0001"
                    },
                    "initial_diameter": {
                        "description": "Initial diameter (mm)",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "final_diameter": {
                        "description": "Final diameter (mm)",
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "workpart_length": {
                        "description": "Workpart length (mm)",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "machining_time": {
                        "description": "Machining time (min)",
                        "type": "integer",
                        "example": 10
                    },
                    "product_quantity": {
                        "description": "Product quantity",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "MachiningProject": {
                "required": [
                    "tool_item_id",
                    "workpiece_material",
                    "machining_process",
                    "cutting_speed",
                    "depth_of_cut",
                    "feeding",
                    "early_tool_life",
                    "is_active"
                ],
                "properties": {
                    "id": {
                        "description": "Machining project ID",
                        "type": "integer",
                        "example": 1
                    },
                    "tool_item_id": {
                        "description": "Tool item ID",
                        "type": "integer",
                        "example": 1
                    },
                    "workpiece_material": {
                        "description": "Workpiece material",
                        "type": "string",
                        "example": "Carbon steel"
                    },
                    "machining_process": {
                        "description": "Machining process",
                        "type": "string",
                        "example": "Turning"
                    },
                    "cutting_speed": {
                        "description": "Cutting speed (m/min)",
                        "type": "integer",
                        "example": 100
                    },
                    "depth_of_cut": {
                        "description": "Depth of cut (mm)",
                        "type": "integer",
                        "example": 50
                    },
                    "feeding": {
                        "description": "Feeding (mm/r)",
                        "type": "integer",
                        "example": 100
                    },
                    "early_tool_life": {
                        "description": "Early tool life (minute)",
                        "type": "integer",
                        "example": 100
                    },
                    "is_active": {
                        "description": "Is active",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Timestamp of creation",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01 00:00:00"
                    },
                    "updated_at": {
                        "description": "Timestamp of last update",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01 00:00:00"
                    }
                },
                "type": "object"
            },
...
            "ToolMaterial": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Tool material ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Tool material name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "description": {
                        "description": "Description of the tool material",
                        "type": "string",
                        "example": "Karbida is a material used for tools"
                    },
                    "created_at": {
                        "description": "Timestamp of creation",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01 00:00:00"
                    },
                    "updated_at": {
                        "description": "Timestamp of last update",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01 00:00:00"
                    },
                    "products": {
                        "description": "Tool products",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ToolProduct"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}